adds the regression targets entries to the GUI and batch (rt_trial saved as empty and does not appear on the interfaces, but will surely be used in a later version).
Batch and GUI compatible for data and design and prepare feature set.

+++ prt_init_fs.m
@@ -269,7 +269,7 @@ else
             PRT.fas(mids(i)).dat = file_array(...
                 datname, ...                 % fname     - filename
                 [szm(i),n_vox],...           % dim       - dimensions (default = [0 0] )
-                spm_type('float32'), ...     % dtype     - datatype   (default = 'float')
+                spm_type('float64'), ...     % dtype     - datatype   (default = 'float')
                 0, ...                       % offset    - offset into file (default = 0)
                 1);                          % scl_slope - scalefactor (default = 1)
         else
